+++ src/theta_implementation.h
@@ -70,7 +70,7 @@ extern struct double_polynomial add_double_polynomials(struct double_polynomial 
 extern struct double_polynomial multiply_double_polynomials(struct double_polynomial P, struct double_polynomial Q);
 extern void print_double_polynomial(struct double_polynomial P, char c);
 extern struct double_polynomial lagrange_product(int max_degree, double* inputs);
-extern struct double_polynomial synthetic_division(struct double_polynomial P, int a);
+extern struct double_polynomial synthetic_division(struct double_polynomial P, double a);
 extern struct double_polynomial lagrange_interpolate(int max_degree, double* inputs, double* outputs);
 extern void adjust_degree(struct double_polynomial* P);
